chore : 주석제거, 콘솔로그 제거
